Update getting-started.md
Updated Step 21 as it was showing the wrong details. The reviewer should be seeing fields filled by the initiator, if we create a new form with new fields it will not see anything.
Remove Step 23. There is no point in adding "Approve" or "Reject" outcomes unless you define different flows for those different outcomes.