Docs should point to correct version number
Hi,

I've posted this documantion suggestion on the official alfresco discord server and as told by @aborroy  I'm therefore opening this ticket:

----

In my acs ansible tasks or in other source code I'm writing for the customer, I'm always referencing to the official documentation pages. 
For example, for an ACS 6.2 ansible task, I'm referencing to
https://docs.alfresco.com/content-services/6.2/install/zip/tomcat/#install-alfresco-wars

For the current ACS 7.2 version I can only reference to:
https://docs.alfresco.com/content-services/latest/install/zip/tomcat/#install-alfresco-wars

What I would like to reference in my sourcecode is:
https://docs.alfresco.com/content-services/7.2/install/zip/tomcat/#install-alfresco-wars
so I don't have to adjust my commits once the next ACS version 7.3 is out...

This Issue is not only related to ACS, it's a general topic for all alfresco services documentations.
Is it somehow possible to upload/reference the latest documentation as well for the specific version number?