[DOCS-6846] Update notes for enterprise users who install ACS with zip (#745)
* Updated the notes for deprecation of legacy transform service

Added additional lines to indicate users to install the required binaries on transform service layer in order for transform service to be fully functional. I have been getting feedback that this part of the notes doesn't mention this important detail and users are confused with the setps.

* Updated notes for community users who install acs with zip and wish to use community transform service

Added additional lines to indicate users to install the required binaries on transform service layer in order for transform service to be fully functional. I have been getting feedback that this part of the notes doesn't mention this important detail and users are confused with the setps.

* Updated notes for enterprise users who install acs with zip

Added additional lines to indicate users to install the required binaries on transform service layer in order for transform service to be fully functional. I have been getting feedback that this part of the notes doesn't mention this important detail and users are confused with the setps.

* Updated notes for enterprise users who install acs with zip

Added additional lines to indicate users to install the required binaries on transform service layer in order for transform service to be fully functional. I have been getting feedback that this part of the notes doesn't mention this important detail and users are confused with the setps.

* Minor changes to the notes on latest version
* Minor changes to the notes on community version
* Minor changes to the notes on 7.1 version
* Minor changes to the notes on 7.0 version