Solr Communication Method Secret with ACS 7.2
In the current documentation says, that in ordner to use the HTTP communication method with "secret" word in request header 
you have to install Alfresco Content Services 7.1 or above.

In Reference to 
https://community.hyland.com/en/blog/posts/83712-upgrading-to-alfresco-content-services-7-2
the "secret" communcation method feature is only available in ACS 7.2 and above.

I know that it was tested initially by @aborroy with ACS7.1, see
https://github.com/AlfrescoLabs/alfresco-solr-comm
and he showed this "secret" communication method feature in
https://www.youtube.com/watch?v=BSNJz6nd8Xg

but as far as we now, this feature is only available for customers with ACS 7.2. and above.

Could you please verify this?