[DOCS-7013] Fix site config for latest DTE 2.3 docs (#864)
If you go to any DTE v2.2 docs page (https://docs.alfresco.com/transformation-engine/2.2/) and then try to switch to the latest docs by selecting "2.3 latest" from the dropdown, it just switches back to showing the earlier 2.2 release docs. This is because the "latest" version is set twice, for 2.2 and 2.3.